+++ client/package.json
@@ -17,6 +17,7 @@
     "updates": "^16.4.2",
     "web-vitals": "^2.1.4"
   },
+  "proxy":"http://localhost:5000",
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
